Begin converting UI to Bootstrap
