+++ public/styles.css
@@ -1,13 +1,5 @@
 html,
 body,
 #app {
-  height: 544px;
-  width: 100vw;
-}
-
-#app {
-  display: flex;
-  flex: 1;
-  font-family: sans-serif;
-  font-weight: 200;
+  height: 100vh;
 }
