+++ src/index.tsx
@@ -3,5 +3,7 @@ import { render } from 'react-dom';
 
 import ReaProjectApp from './components/ReaProjectApp';
 
+import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
+
 const rootElement = document.getElementById('app');
 render(<ReaProjectApp />, rootElement);
